No throw when union doesn't contain field values
`makeUnion` function contains a shortcut when the `values` array is empty. However, this prevents the code from going to `mixedArray` which has a safeguard when the `values` array doesn't correspond to the number of union case fields. So if `values = [||]` but the union case has one or more fields, the code will throw instead of showing a typed error. Example:

```fsharp
type Foo = Foo of int

// Note the JSON is an array but doesn't contain the field value
let foo = Decode.Auto.fromString("[\"Foo\"]")
```