Extend the DateTime decoding API
The current API has the behavior of decoding a given time, then converting it to
UTC. This leads to confusing behavior, e.g. in unit tests when a given string
doesn't parse to the same datetime it represents. There's some discussion of
this in [issue 107](https://github.com/thoth-org/Thoth.Json/issues/107).

I've elected to try adjusting this with two methods discussed in #107. This
deprecates the `datetime` decoder, and adds two new decoders, `datetimeUTC` and
`datetimeLocal`.

I've tried to update the tests, but -- and this is silly -- I can't seem to get
them to _run_. The `tests` project is written as a library, not an executable,
so running `dotnet test` does nothing, and it cannot be `dotnet run`. `fake-cli`
is in the tool manifest, but there's no build script. The Azure CI configs call
a script called `fake.sh`, which doesn't exist in the project. I can't find a
`CONTRIBUTING.md` or similar, so I'm just... a bit lost. If a maintainer can let
me know how the tests work, I'm happy to run them, make sure they pass.

Thanks for the great project! Thoth is a pleasure to use.

Closes #107.